Backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
alembic
psycopg2-binary
python-dotenv
requests
httpx
scikit-learn
plaid-python
transformers
torch
